lowlight-enhancer-uv/pyproject.toml
[project]
name = "lowlight-enhancer-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.0",
    "flask>=3.1.2",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
]
